requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
langchain>=0.0.335
langchain-core>=0.1.23
langchain-groq>=0.0.1
python-dotenv==1.0.0
pydantic>=1.10.0,<2.0.0
jinja2>=3.0.0
pyspellchecker==0.7.2